assem-vc_paddle/requirements.txt
g2p-en==2.1.0
inflect==5.3.0
librosa==0.8.1
omegaconf==2.0.0
unidecode==1.1.1
paddlepaddle-gpu==1.4.0
tensorboard==2.2.0
pypaddlepaddle-gpu
pyyaml==5.3.1
soundfile==0.10.3.post1
matplotlib==3.1.3
pysptk==0.1.18
xpinyin==0.5.7
